Major cleanup
Changed name of CD_ConastraintDiagnostics to CD_ConstraintViewer. Removed unused functions from code. Cleaned up space errors
